+++ changelog.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 # [Unreleased](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v8.2.0...HEAD)
+* [PCM-2220](https://inindca.atlassian.net/browse/PCM-2220) - Automatically accept a pendingSession that comes in after the "fake" pendingSession was answered
+
 # [v8.2.0](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v8.1.6...v8.2.0)
 * [PCM-2213](https://inindca.atlassian.net/browse/PCM-2213) - Allow dynamic adding and removing of supported session types